Add a check for stuck head exception in step
Stepping to track 0 is a special case in greaseweazle firmware. It
steps down 256 times or until TRACK0 is asserted, whichever is less.
If TRACK0 is not asserted at the end of the procedure it returns
Ack.NoTrk0. This means we can't get seek errors when stepping to
ecyl, but the function is re-used to prevent code duplication.

Initially seeking to track 0 is not required - the hardware does
this when necesssary.