gw convert: Iterate over output tracks, processing input as required.
This more natural and potentially more efficient, as we will not
process input tracks we do not need for the output. A further benefit
is that slow emit_track() calls (such as SCP) will be intyerleaved
with input processing so that execution time gets naturally accounted
to the track-by-track input reporting.