score_port: Accept "gw-compat" products
If the USB device's product string contains "gw-compat" (case
insensitive), it will be treated as a device greaseweazle can use,
albeit at a lower priority than a genuine Greaseweazle.

Together with a 1-liner change in Adafruit_Floppy's compatible firmware,
this allows `gw` commands to work with that firmware without specifying
a --device to every command.